SA/requirements.txt
Flask
python-dotenv
praw
pandas
transformers
torch